backend/requirements.txt
fastapi==0.128.7
uvicorn==0.40.0
mariadb==1.1.14
python-dotenv==1.2.1